wired up proposal handler
