+++ x/stakeibc/client/proposal_handler.go
@@ -8,4 +8,5 @@ import (
 
 var (
 	AddValidatorsProposalHandler = govclient.NewProposalHandler(cli.CmdAddValidatorsProposal)
+	ToggleLSMProposalHandler     = govclient.NewProposalHandler(cli.CmdToggleLSMProposal)
 )
