+++ x/stakeibc/module.go
@@ -131,7 +131,7 @@ func (am AppModule) Name() string {
 
 // Route returns the capability module's message routing key.
 func (am AppModule) Route() sdk.Route {
-	return sdk.NewRoute(types.RouterKey, NewHandler(am.keeper))
+	return sdk.NewRoute(types.RouterKey, NewMessageHandler(am.keeper))
 }
 
 // QuerierRoute returns the capability module's query routing key.
